chat/requirements.txt
llama-index-core
llama-index-embeddings-openai
llama-index-llms-groq
fastapi
uvicorn
requests